backend/requirements.txt
# ── Core framework ────────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2
asyncpg==0.29.0

# ── Pydantic ──────────────────────────────────────────────────────────────────
pydantic[email]==2.8.2
pydantic-settings==2.3.4

# ── AI / OpenAI ───────────────────────────────────────────────────────────────
openai==1.35.14

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Cache / PubSub ────────────────────────────────────────────────────────────
redis[asyncio]==5.0.7

# ── Rate limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.1
aiosmtplib==3.0.1
